+++ src/main/java/org/grnet/pidmr/dto/UpdateProviderV2.java
@@ -1,0 +1,66 @@
+package org.grnet.pidmr.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.Valid;
+import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
+import java.util.HashSet;
+import java.util.Set;
+
+@Schema(name="UpdateProvider", description="An object represents a request for updating a Provider.")
+public class UpdateProviderV2 {
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "Type of PID.",
+            example = "ark"
+    )
+    @JsonProperty("type")
+    public String type;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "Provider name.",
+            example = "ARK alliance"
+    )
+    @JsonProperty("name")
+    public String name;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "Short Provider description.",
+            example = "Archival Resource Keys (ARKs) serve as persistent identifiers, or stable, trusted references for information objects."
+    )
+    @JsonProperty("description")
+    public String description;
+
+    @Schema(
+            type = SchemaType.ARRAY,
+            implementation = ResolutionModeRequest.class,
+            description = "The resolution modes supported by Provider."
+    )
+    @JsonProperty("resolution_modes")
+    public Set<@Valid ResolutionModeRequest> actions = new HashSet<>();
+
+    @Schema(
+            type = SchemaType.ARRAY,
+            implementation = String.class,
+            description = "The regexes supported by Provider.",
+            example = "[\"^(a|A)(r|R)(k|K):(?:/d{5,9})+/[a-zA-Zd]+(-[a-zA-Zd]+)*$.\"]"
+    )
+    @JsonProperty("regexes")
+    public Set<String> regexes = new HashSet<>();
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "A PID example.",
+            example = "ark:/13030/tf5p30086k"
+    )
+    @JsonProperty("example")
+    public String example;
+}