+++ src/main/java/org/grnet/pidmr/endpoint/AdminEndpoint.java
@@ -37,7 +37,7 @@ import org.grnet.pidmr.dto.AdminProviderDto;
 import org.grnet.pidmr.dto.InformativeResponse;
 import org.grnet.pidmr.dto.ProviderDto;
 import org.grnet.pidmr.dto.ProviderRequest;
-import org.grnet.pidmr.dto.UpdateProviderDto;
+import org.grnet.pidmr.dto.UpdateProvider;
 import org.grnet.pidmr.dto.UpdateProviderStatus;
 import org.grnet.pidmr.enums.ProviderStatus;
 import org.grnet.pidmr.exception.ConflictException;
@@ -295,7 +295,7 @@ public class AdminEndpoint {
             example = "1",
             schema = @Schema(type = SchemaType.NUMBER)) @PathParam("id")
                            @Valid @NotFoundEntity(repository = ProviderRepository.class, message = "There is no Provider with the following id:") Long id,
-                           @Valid @NotNull(message = "The request body is empty.") UpdateProviderDto request) {
+                           @Valid @NotNull(message = "The request body is empty.") UpdateProvider request) {
 
         ProviderDto response = null;
         try {