+++ CHANGELOG.md
@@ -24,6 +24,11 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 ---
 
+### Added
+
+-   [#59](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/59) - PIDMR-146 Add "endpoint" field to Provider entity with mode-specific endpoints.
+
+
 ## 2.1.0 - 2024-04-29
 
 ---